Move RadioOption to its own file
