+++ packages/components/src/RadioGroup/index.ts
@@ -1,1 +1,2 @@
-export { RadioGroup, RadioOption } from "./RadioGroup";
+export { RadioGroup } from "./RadioGroup";
+export { RadioOption } from "./RadioOption";
